refactored tasks
