+++ src/Payload/Family/FamilyPayload.php
@@ -18,9 +18,6 @@ final class FamilyPayload extends AbstractPayload
     /** @var \Akeneo\Pim\ApiClient\Pagination\ResourceCursorInterface|null */
     private $resources;
 
-    /** @var \Akeneo\Pim\ApiClient\Pagination\ResourceCursorInterface|null */
-    private $modelResources;
-
     public function __construct(
         AkeneoPimEnterpriseClientInterface $akeneoPimClient,
         ?CommandContextInterface $commandContext = null
@@ -39,17 +36,5 @@ final class FamilyPayload extends AbstractPayload
     public function setResources(ResourceCursorInterface $resources): void
     {
         $this->resources = $resources;
-    }
-
-    public function getModelResources(): ?ResourceCursorInterface
-    {
-        return $this->modelResources;
-    }
-
-    public function setModelResources(?ResourceCursorInterface $modelResources): self
-    {
-        $this->modelResources = $modelResources;
-
-        return $this;
     }
 }
