+++ src/Task/ProductModel/ProcessProductModelsTask.php
@@ -13,9 +13,9 @@ use Synolia\SyliusAkeneoPlugin\Logger\Messages;
 use Synolia\SyliusAkeneoPlugin\Payload\PipelinePayloadInterface;
 use Synolia\SyliusAkeneoPlugin\Payload\ProductModel\ProductModelPayload;
 use Synolia\SyliusAkeneoPlugin\Provider\ConfigurationProvider;
-use Synolia\SyliusAkeneoPlugin\Task\AbstractBatchTask;
+use Synolia\SyliusAkeneoPlugin\Task\AbstractProcessTask;
 
-final class ProcessProductModelsTask extends AbstractBatchTask
+final class ProcessProductModelsTask extends AbstractProcessTask
 {
     /** @var \Synolia\SyliusAkeneoPlugin\Provider\ConfigurationProvider */
     private $configurationProvider;
@@ -60,6 +60,7 @@ final class ProcessProductModelsTask extends AbstractBatchTask
         );
 
         $this->handle($payload, $resources);
+        $this->processManager->waitForAllProcesses();
 
         return $payload;
     }
